[REF] cetmix_tower_server: Variable values
Before this commit
------------------

Variable values were obtained by calling the dedicated function implemented in
the `cx.tower.variable.mixin` model.
This was done to unify this function by making it available directly in the
models that are inheriting this mixin.
This worked well when Server was the only model, however with the addition of
Jets and Jet Templates, this approach was not working anymore.
Because it required a complex logic that called all three models separately
and merged the results together later.

After this commit
------------------

Variable values are now obtained by calling the dedicated function implemented in
the `cx.tower.variable` model.
This function accepts a list of variable references and together with the server,
jet template and jet records.
This allows to fetch all the variable values in a single call while respecting
the order those variable values are applied.
It also makes the code more readable and maintainable.